test(increment4): consolidate final authority regressions
